Added A Filter to show only the root level of an Activity View. Activating this filter hides all sub-activity diagrams. In practice, it just hides all DecomposeSubNode Nodes.
Added the ability to open FlowDiagram-linked elements in a separate ERS/AD View Diagram.
That way a user can focus on a specific sub-part of a complex diagram.
It also allows the user to quickly navigate between ERS View and Activity Diagram View.

To use it -> Right click on a graphical element linked to a FlowDiagram